third_party/WebKit/LayoutTests/fast/dom/StyleSheet/gc-parent-rule-expected.txt

Test that CSS rules don't lose custom properties on GC when only reachable through child rules.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS style.sheet is an instance of CSSStyleSheet
PASS style.sheet is null
FAIL rule.parentRule.foo should be bar. Threw exception TypeError: Cannot read property 'foo' of null
PASS successfullyParsed is true

TEST COMPLETE
